PAY-65: Put print_response function in its own file and autoload it with composer
